moved report generator to seperate file and modified cron pattern
